requirements.txt
Flask>=2.3.0,<3.0
python-docx>=0.8.11,<1.0
python-pptx>=0.6.21,<1.0
Pillow>=9.0.0,<10.0
requests>=2.28.0,<3.0
python-dotenv>=0.21.0,<1.0
openai>=0.27.0,<1.0
PyYAML>=6.0,<7.0